+++ src/presentation/handler/topics_handler.py
@@ -5,15 +5,13 @@ from src.core.logs import info, debug, error
 preview_manager = PreviewManager()
 
 
-def get_topics_from_app(item_name, *, page: int, per_page: int, kinds: list[str] | None, time_window: str | None):
+def get_topics_from_app(item_name, *, page: int, per_page: int):
 
 
     
     api_data = preview_manager.get_topics(item_name=item_name,
         page=page,
-        per_page=per_page,
-        kinds=kinds,
-        time_window=time_window,
+        per_page=per_page
     )
     
     return api_data